requirements.txt
boltons==25.0.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
easydict==1.13
filelock==3.12.2
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.7.0
joblib==1.3.2
numpy==1.21.6
opencv-python==4.10.0.84
packaging==24.0
pandas==1.3.5
patsy==1.0.1
Pillow==9.3.0
pyaml==23.5.8
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
requests==2.28.1
scikit-learn==1.0.2
scikit-optimize==0.10.2
scipy==1.7.3
six==1.17.0
statsmodels==0.13.5
threadpoolctl==3.1.0
timm==0.6.12
torch==1.12.1+cu116
torchdiffeq==0.2.5
torchsde==0.2.5
torchvision==0.13.1+cu116
tqdm==4.67.1
trampoline==0.1.2
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.15.0
